Rename package files
